+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 - Turn into a Nix Flake.
 - Improve serializer to be more robust in end-of-stream conditions.
 - Bump requirements to require at least GCC version 9.
+- Internally use `stream` instead of `tcp` and `datagram` for `udp`.
 
 ### Removed
 - Badges (eg. LGTM and build status) in README and Hydra.